Remove images from Grove blocks per issues/49.
Just a suggestion - up to you if you want to merge it.

Removing the images makes them about a third the size, which makes it possible to fit more blocks on screen, as well as making the program a lot more streamlined.

I don't think the images add a great deal to readability - anyone with a module will know what it's called, and a thumbnail sized image will probably not help much.